ask_hr_ai_agent_adk/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
babel==2.17.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
docxcompose==1.4.0
docxtpl==0.16.7
Flask==3.1.2
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-genai==1.55.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
outcome==1.3.0.post0
packaging==25.0
proto-plus==1.27.0
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
PySocks==1.7.1
python-docx==1.1.2
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
selenium==4.39.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tenacity==9.1.2
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.4
wsproto==1.3.2
